backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
groq==0.13.0
httpx==0.27.2
email-validator==2.2.0
mailjet-rest==1.3.4
vonage==3.11.0
vonage-sms==1.1.0
